+++ packages/continuum/CHANGELOG.md
@@ -1,4 +1,4 @@
-## UNRELEASED
+## 2.0.0 - 2026-01-08
 
 ### Breaking Changes